usb: gadget: adb: Only enable the gadget when adbd is ready
When adb is enabled, only connect the gadget when adbd is ready. If adbd
dies or is restarted (e.g. "adb root"), the gadget is disconnected when
the adb device is close, and it is re-connected once adb re-open the
device.

- Add callbacks to adb, similar to FunctionFs callbacks, to notify the
gadget when the daemon is ready or closed.
- Refcount calls to android_enable/android_disable to enable the gadget
only once all the function daemons are ready.
- Add enable/disble to android_usb_function to notify the function when
it is added/removed from the list of enabled functions.

Change-Id: Ibd803a64da9ce553e64a66b2991b66908975c9e6
Reviewed-on: http://mcrd1-5.corpnet.asus/code-review/master/47121
Reviewed-by: Yi-Hsin Hung <Yi-Hsin_Hung@asus.com>
Tested-by: Yi-Hsin Hung <Yi-Hsin_Hung@asus.com>
Reviewed-by: Leslie Yu <Leslie_Yu@asus.com>
Reviewed-on: http://mcrd1-5.corpnet.asus/code-review/master/48103
Tested-by: Jive Hwang <jive_hwang@asus.com>
Reviewed-by: Jive Hwang <jive_hwang@asus.com>